Move struct stkmon_save_s to apps/include/system/utils.h
